Support options from task and config
Previously, if you attempted to mix options from the config file or
via task switches, they would clobber each other producing mixed results.
But it wasn't clear this behavior was not supported.

This changes to support reading from config file _and_ mix task switches
merging the options together. The task switches take precedence.

It also defaults to using the config file if it is present, but leaves
the switch so that `--no-config` could be used to explicitly ignore
the config file instead